テストファイルを適切な場所に移動（tests/components/BarcodeScanner.test.tsx → tests/hooks/useBookScanner.test.tsx）
